% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{ntnu_pal}
\alias{ntnu_pal}
\title{NTNU palette with ramped colours}
\usage{
ntnu_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{List of palettes to choose from}

\item{reverse}{If TRUE, reverses colour direction}

\item{...}{Additional arguments}
}
\description{
NTNU palette with ramped colours
}
\examples{
\dontrun{
 ntnu_pal()
}
}
