% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ntnu.R
\name{theme_ntnu}
\alias{theme_ntnu}
\title{Add NTNU theme to ggplot chart}
\usage{
theme_ntnu(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{Basis for text and line sizes}

\item{base_family}{Base font family}
}
\description{
Add NTNU theme to ggplot chart
}
\examples{
line <- ggplot2::ggplot(line_df, aes(x = year, y = lifeExp)) +
ggplot2::geom_line(colour = "#007f7f", size = 1) +
ggplot2::geom_hline(yintercept = 0, size = 1, colour="#333333") +
theme_ntnu()
}
