% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_ntnu}
\alias{scale_color_ntnu}
\title{Setup colour palette for ggplot2}
\usage{
scale_color_ntnu(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Palette to apply}

\item{discrete}{Whether the scale is continuous or discrete}

\item{reverse}{Whether the palette should be reversed}

\item{...}{Additional arguments}
}
\description{
Setup colour palette for ggplot2
}
\examples{
\dontrun{
scale_color_ntnu()
}
}
