% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_ntnu}
\alias{scale_fill_ntnu}
\title{Setup fill palette for ggplot2}
\usage{
scale_fill_ntnu(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Palette to apply}

\item{discrete}{Whether the scale is continuous or discrete}

\item{reverse}{Whether the palette should be reversed}

\item{...}{Additional arguments}
}
\description{
Setup fill palette for ggplot2
}
\examples{
\dontrun{
scale_fill_ntnu()
}
}
